integrations/calling/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
twilio==9.0.0
google-generativeai==0.3.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.12.5
tabulate==0.9.0
gTTS==2.5.4
python-multipart==0.0.22
